I discovered new flaws in the source files!
Unable to resolve a nagging issue, I decided to run the HTML and CSS files through an online validator. The HTML came back clean, but the CSS was flagged for errors. Once I corrected those errors, the issue was resolved. I'm going to re-create the new-header branch and start again.