First checkpoint
I've almost got my video background working, but in the process I inadvertently deleted a section of code from the HTML file. That's what I get for opening that file in two editors! :-( I'm going to switch back to the master branch and retrieve that version of the HTML file. I'll pull out the code that I need and then switch back to the new-header branch.