Putting the finishing touches on the video background.
I believe I've got the video background working the way I want. Additionally, I made some general refinements and enhancements. I try to limit the changes I make when working on a branch, but sometimes I get carried away. When I merge this with the master, conflict resolution is going to be a bear!