Yikes!
Hopefully I've avoided a disaster. I got carried away modifying the HTML and CSS files in a separate branch. Unfortunately, I'd made too many changes for Git to resolve on its own. So, I had to pull the files from the master as well as the files from the branch and compare them with an external program. Fortunately, I have one that's pretty good. It's called "Beyond Compare."