I think I've got it working!
The port of the code I developed in CodePen did not go smoothly. For reasons unknown, it did not produce the desired result. Instead, I had to fiddle with it for a couple of hours before I got it working. I made wholesale changes to the HTML file, and expect the impending merge to be a nightmare!