Layout complete
All of the elements have been laid out as close to the spec as I could get. For some reason, the video in the last section is no longer aligned with the top of its container. Instead, it's vertically centered. I'm not sure when or why that changed. Perhaps merging this branch with the master will fix that.