Conflict resolution
The changes to index.html and style.css reflect conflicts I resolved outside of the Visual Studio Code environment.

This has been a long time in coming. Except for tweaks to the typefaces and colors, I believe I have a pre-deployment product. I'll create a separate branch to work on the responsive design aspects.