A number of modifications
In the header, I added an inverted image to the right side of "Colmar Academy." In the countdown timer pop-up, I reconstructed the close button using a div element instead of a button element. I did so because I was having difficulty centering the text (X) in the <button> element. I'm not familiar enough with the <button> tag to know what it's default properties are, and whether any of those might be interfering with the positioning of the text.