Redesigned the University's logo.
I redesigned the university's logo. The new design conforms to my original concept for that element. I think it's cleaner and has a more contemporary look.

I disabled the diffuser for the background video so that the latter will play at full saturation and intensity. Once the video stops, the diffuser will be activated to reduce the background's intensity by 50 percent.

It's probably not a good idea to address two issues in one branch. This time around, however, I might get away with it. I'm 99.9% sure I'll keep both changes.