Smooth scrolling successfully implemented.
Smooth scrolling is a feature I like to implement in all of my HTML documents. Although I've done so in this one, it's with limited effect. Smooth scrolling controls the motion of the document when jumping between points within the document. The original design specification does not call for such links, and I didn't add any. The only time the effect is in evidence is when clicking on the logo to return to the top of the page. Also, I rebuilt the close button on the countdown timer. As a result, the cross members of the "X" are perpendicular to each other, and the color of the timer is beige with a blue border.