Whew! That was close!
When I discovered that elements were being displayed that should only have been displayed when the condition of the media query was met, I poured over the code looking for the culprit. Experience is shown that even a single character out of place can ruin everything. Eventually I discovered that I had forgotten to insert display commands to ensure that the conditional elements would, by default, be hidden.